requirements.txt
amqp==2.1.4
anyjson==0.3.3
Babel==2.3.4
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
billiard==3.5.0.2
celery==4.0.2
certifi==2015.4.28
contextlib2==0.5.4
coverage==3.5
Django==1.10.3
django-braces==1.10.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-filter==0.15.3
django-http-proxy==0.4.3
django-import-export==0.5.1
django-jsonfield==1.0.1
django-mathfilters==0.3.0
django-oauth-toolkit==0.10.0
django-prometheus==1.0.8
django-qiniu-storage==2.3.0
django-redis-cache==1.7.1
django-statsd-mozilla==0.3.16
django-tagging==0.4.3
djangorestframework==3.5.2
dogslow==0.9.7
drf-extensions==0.3.1
flower==0.9.1
futures==3.0.5
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
hiredis==0.2.0
html5lib==0.90
httplib2==0.9
imageio==1.6
kombu==4.0.2
lxml==2.3
Markdown==2.6.6
MyQR==2.0.1
mysqlclient==1.3.10
MySQL-python==1.2.5
nose==1.3.1
numpy==1.11.0
oauth2-provider==0.0
oauthlib==1.0.3
Pillow==3.3.1
Pillow-PIL==0.1.dev0
pingpp==2.0.1
poster==0.8.1
prometheus-client==0.0.18
pyasn1==0.1.9
pyExcelerator==0.6.4.1
pymongo==3.3.0
PyMySQL==0.7.9
PyPDF2==1.19
python-dateutil==2.5.0
python-memcached==1.53
pytz==2016.10
qiniu==7.0.5
qrcode==5.1
raven==5.31.0
redis==2.10.5
reportlab==2.5
requests==2.9.1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
South==1.0
SQLAlchemy==1.1.9
sqlparse==0.2.2
statsd==2.1.2
tblib==1.3.0
tornado==4.2
vine==1.1.3
wheel==0.24.0
xhtml2pdf==0.0.6
XlsxWriter==0.8.4
pycrypto==2.6.1
django-admin-view-permission==0.4
python-keyczar==0.716
django-encrypted-fields==1.1.2

# sudo apt-get install libcurl4-openssl-dev
# pycurl==7.4.3
##################### Debug Requirement #########################
django-debug-toolbar==1.6


###################### Extras Requirement #########################
# apt-get update;apt-get install -y gcc libxml2-dev libxslt1-dev python-dev libmysqld-dev python-pip supervisor git
# sudo apt-get install libtiff4-dev libjpeg-dev libjpeg8-dev zlib1g-dev libfreetype6-dev liblcms1-dev libwebp-dev
# pip install -i http://pypi.oneapm.com/simple --upgrade blueware
# pip install -i http://pypi.oneapm.com/simple --upgrade oneapm-ci-sdk
# pip install git+git://github.com/xmeron/django-prometheus.git
